Extract, and rename, the fake activity monitor
